+++ changelog.md
@@ -10,6 +10,7 @@ Tentative: 2023-04-04
 * Add sign-in warning of uBlock Origin / ORCID issue
 * Add "Copy handle" button to the home page list
 * Add download for "recently minted raids" report
+* Add subjects, related Raids, related objects and alternate identifiers to metadata
 
 
 # 1.0